added simple pkg and move some code to it
